# backend/openapi.json
{"openapi":"3.0.0","info":{"title":"Vending machine","description":"Coding challenge","version":"1.0.0"},"servers":[{"url":"http://127.0.0.1:3000/api"}],"components":{"securitySchemes":{"cookieAuth":{"type":"apiKey","in":"cookie","name":"vmsession"}},"responses":{"BadRequest":{"description":"Validation error","content":{"application/json":{"schema":{"type":"object","properties":{"fieldErrors":{"type":"array","description":"Error messages for form fields","items":{"$ref":"#/components/schemas/FieldError"}},"nonfieldErrors":{"type":"array","description":"General error messages list","items":{"type":"string"}}}}}}}},"schemas":{"FieldError":{"type":"object","properties":{"field":{"type":"string","description":"Field name"},"message":{"type":"string","description":"Error message"}}},"LoginRequest":{"type":"object","required":["username","password"],"properties":{"username":{"type":"string","description":"User login"},"password":{"type":"string","description":"User password"}}},"UserUpdateRequest":{"type":"object","required":["username"],"properties":{"username":{"type":"string","description":"User login"},"password":{"type":"string","description":"User password"}}},"SignupRequest":{"type":"object","required":["username","role","password"],"properties":{"username":{"type":"string","maxLength":128,"description":"User login"},"role":{"type":"integer","description":"User role (1 - seller, 2 - buyer)","enum":[1,2]},"password":{"type":"string","maxLength":128,"description":"User password"}}},"MessageResponse":{"type":"object","properties":{"message":{"type":"string"}}},"User":{"type":"object","properties":{"id":{"type":"integer","format":"int64","description":"User id"},"username":{"type":"string","maxLength":128,"description":"User login"},"role":{"type":"integer","description":"User role (1 - seller, 2 - buyer)","enum":[1,2]},"deposit":{"type":"integer","description":"Current user deposit"}}},"ProductBody":{"type":"object","required":["productName","amountAvailable","cost"],"properties":{"productName":{"type":"string","maxLength":128,"description":"Product name"},"amountAvailable":{"type":"integer","description":"Amount available"},"cost":{"type":"integer","description":"Product price"}}},"Product":{"type":"object","properties":{"id":{"type":"integer","format":"int64","description":"Product id"},"productName":{"type":"string","maxLength":128,"description":"Product name"},"amountAvailable":{"type":"integer","description":"Amount available"},"cost":{"type":"integer","description":"Product price"}}},"BuyBody":{"type":"object","required":["productId","quantity"],"properties":{"productId":{"type":"integer","format":"int64","description":"Product id"},"quantity":{"type":"integer","description":"Quantity of products ordered"}}},"DepositBody":{"type":"object","required":["value"],"properties":{"value":{"type":"integer","enum":[5,10,20,50,100],"description":"Coin denomination"}}},"BuyResponse":{"type":"object","properties":{"productId":{"type":"integer","format":"int64","description":"Product id"},"total":{"type":"integer","description":"Total price"},"quantity":{"type":"integer","description":"Quantity of products ordered"},"change":{"type":"array","items":{"type":"integer"},"description":"Coin change"}}}}},"security":[{"cookieAuth":[]}],"paths":{"/auth/login":{"post":{"summary":"Logs in and returns the authentication cookie and user data","tags":["auth"],"requestBody":{"required":true,"description":"A JSON object containing the login and password.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}}},"security":[],"responses":{"200":{"description":"Successfully authenticated. The session ID is returned in a cookie named `vmsession`. You need to include this cookie in subsequent requests.\n","headers":{"Set-Cookie":{"schema":{"type":"string","example":"vmsession=abcde12345; Path=/; HttpOnly"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"$ref":"#/components/responses/BadRequest"}}}},"/auth/logout":{"post":{"summary":"Logs out (removes current session)","tags":["auth"],"responses":{"200":{"description":"Successfully logged out.\n","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}}}}},"/auth/logout/all":{"post":{"summary":"Logs out from all devices (removes all current user's sessions)","tags":["auth"],"responses":{"200":{"description":"Successfully logged out.\n","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}}}}},"/auth/me":{"post":{"summary":"Gets current user info","tags":["auth"],"responses":{"200":{"description":"Successful operation.\n","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/products":{"get":{"tags":["product"],"summary":"Returns all products","responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}}}},"post":{"summary":"Adds new product","tags":["product"],"requestBody":{"required":true,"description":"A JSON object containing product info.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductBody"}}}},"responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"400":{"$ref":"#/components/responses/BadRequest"}}}},"/products/{productId}":{"get":{"tags":["product"],"summary":"Find product by ID","description":"Returns a single product","parameters":[{"name":"productId","in":"path","description":"ID of product to return","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"404":{"description":"Product not found"}}},"put":{"summary":"Updates product data","tags":["product"],"parameters":[{"name":"productId","in":"path","description":"ID of product to update","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"required":true,"description":"A JSON object containing product info.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductBody"}}}},"responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"400":{"$ref":"#/components/responses/BadRequest"},"403":{"description":"Only seller can perform this operation."}}},"delete":{"tags":["product"],"summary":"Deletes a product","description":"Available only to seller who owns it","parameters":[{"name":"productId","in":"path","description":"ID of product to delete","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}}},"403":{"description":"Not owned by seller"},"404":{"description":"Product not found"}}}},"/users":{"post":{"summary":"Registers new user account","tags":["user"],"requestBody":{"required":true,"description":"A JSON object containing the login, password and role.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}}},"security":[],"responses":{"200":{"description":"Successfully created. After signing up user has to sign in.\n","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"$ref":"#/components/responses/BadRequest"}}}},"/users/{userId}":{"put":{"summary":"Updates account info (username/password)","description":"Update limited to username/password only for account owned by user","tags":["user"],"parameters":[{"name":"userId","in":"path","description":"ID of user account to update","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"required":true,"description":"A JSON object containing product info.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdateRequest"}}}},"responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"$ref":"#/components/responses/BadRequest"}}},"delete":{"tags":["user"],"summary":"Deletes user account","description":"Available only to user who owns it","parameters":[{"name":"userId","in":"path","description":"ID of a user account to delete","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"403":{"description":"Not owned by user"},"404":{"description":"User not found"}}}},"/deposit":{"post":{"summary":"Adds coin to deposit","tags":["machine"],"requestBody":{"required":true,"description":"A JSON object coin value.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepositBody"}}}},"responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"$ref":"#/components/responses/BadRequest"},"403":{"description":"Only buyer can perform this operation."}}}},"/buy":{"post":{"summary":"Buys a product","tags":["machine"],"requestBody":{"required":true,"description":"A JSON object with product ID and quantity.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuyBody"}}}},"responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuyResponse"}}}},"400":{"$ref":"#/components/responses/BadRequest"},"403":{"description":"Only buyer can perform this operation."}}}},"/reset":{"post":{"summary":"Resets user's deposit","tags":["machine"],"responses":{"200":{"description":"successful operation","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"400":{"$ref":"#/components/responses/BadRequest"},"403":{"description":"Only buyer can perform this operation."}}}}},"tags":[{"name":"auth","description":"Authentication"},{"name":"user","description":"User account"},{"name":"product","description":"Product"},{"name":"machine","description":"Vending machine operations"}]}